Replace old storage utils file
